+++ routers/index.js
@@ -35,6 +35,7 @@ module.exports = _ => {
   _.get("/portal/home", authentication, portal.home); // *页面跳转 -> 首页
   _.get("/portal/platform", authentication, portal.platform); // *页面跳转 -> 平台
   _.get("/portal/product", authentication, portal.product); // *页面跳转 -> 产品
+  _.get("/portal/workflow/:id", portal.info); // *页面跳转 -> 工作流详情页
   _.get("/portal/directory", authentication, portal.directory); // *页面跳转 -> 公共目录
 
   /*